Only return integers for pages (#2409)

+++ src/WithPagination.php
@@ -18,7 +18,7 @@ trait WithPagination
         $this->page = $this->resolvePage();
 
         Paginator::currentPageResolver(function () {
-            return $this->page;
+            return (int) $this->page;
         });
 
         Paginator::defaultView($this->paginationView());
@@ -53,6 +53,6 @@ trait WithPagination
     {
         // The "page" query string item should only be available
         // from within the original component mount run.
-        return request()->query('page', $this->page);
+        return (int) request()->query('page', $this->page);
     }
 }
